Removed redundant method for ant position tuple size. Completed getter and setter logic and validation for python Ant

+++ c/anttype.h
@@ -28,7 +28,7 @@ struct ant_t
 	ant_orientation_t orientation;
 	ant_position_t position;
 	size_t tuple_size;
-	int pos[];
+	long long int pos[];
 
 };
